Task-3/requirements.txt
streamlit
numpy
faiss-cpu
spacy
scispacy
sentence_transformers